ticket-35/page-decouvertes-recherche-1 : create specific widget for search item categories
